Changed the DestructiveOperations from a static helper to a stateful helper class.
