+++ src/main/java/org/elasticsearch/action/support/DestructiveOperations.java
@@ -20,23 +20,36 @@
 package org.elasticsearch.action.support;
 
 import org.elasticsearch.ElasticsearchIllegalArgumentException;
+import org.elasticsearch.common.logging.ESLogger;
+import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.node.settings.NodeSettingsService;
 
 /**
- * Static helper for dealing with destructive operations and wildcards.
+ * Helper for dealing with destructive operations and wildcard usage.
  */
-public final class DestructiveOperations {
+public final class DestructiveOperations implements NodeSettingsService.Listener {
 
+    /**
+     * Setting which controls whether wildcard usage (*, prefix*, _all) is allowed.
+     */
     public static final String REQUIRES_NAME = "action.destructive_requires_name";
 
-    private DestructiveOperations() {
+    private final ESLogger logger;
+    private volatile boolean destructiveRequiresName;
+
+    // TODO: Turn into a component that be reused and wired up into all the transport actions where
+    // this helper logic is required. Note: also added the logger as argument, otherwise the same log
+    // statement is printed several times, this can removed once this becomes a component.
+    public DestructiveOperations(ESLogger logger, Settings settings, NodeSettingsService nodeSettingsService) {
+        this.logger = logger;
+        destructiveRequiresName = settings.getAsBoolean(DestructiveOperations.REQUIRES_NAME, false);
+        nodeSettingsService.addListener(this);
     }
 
     /**
      * Fail if there is wildcard usage in indices and the named is required for destructive operations.
-     *
-     * @param destructiveRequiresName Controls whether wildcard usage (*, prefix*, _all) is allowed
      */
-    public static void failDestructive(String[] aliasesOrIndices, boolean destructiveRequiresName) {
+    public void failDestructive(String[] aliasesOrIndices) {
         if (!destructiveRequiresName) {
             return;
         }
@@ -53,6 +66,15 @@ public final class DestructiveOperations {
                     throw new ElasticsearchIllegalArgumentException("Wildcard expressions or all indices are not allowed");
                 }
             }
+        }
+    }
+
+    @Override
+    public void onRefreshSettings(Settings settings) {
+        boolean newValue = settings.getAsBoolean("action.destructive_requires_name", destructiveRequiresName);
+        if (destructiveRequiresName != newValue) {
+            logger.info("updating [action.operate_all_indices] from [{}] to [{}]", destructiveRequiresName, newValue);
+            this.destructiveRequiresName = newValue;
         }
     }
